refactor: . moved Link class to own file
